Adding updated logic

+++ Testing.Client/DataAccess/IVolunteerRepo.cs
@@ -5,7 +5,7 @@ namespace Testing.Client.DataAccess
     public interface IVolunteerRepo
     {
         List<Volunteer> GetVolunteers();
-        Volunteer GetVolunteerByPhoneNumber(int phoneNumber);
+        Volunteer GetVolunteerByPhoneNumber(long phoneNumber);
         void SaveVolunteer(Volunteer volunteer);
 
     }
